[Refactoring] SessionId 기반의 안전한 세션 관리 아키텍처 도입 (Phase 2)
- 세션 포인터 제거: Player, GameEvents, PlayerFactory 등 도메인/로직 레이어에서 ISession* 직접 참조를 제거하고 sessionId로 대체
- 안전한 비동기 전송 도입: IDispatcher::WithSession 인터페이스 추가 및 DispatcherImpl 내 세션 레지스트리 구현
- Room 레이어 개선: Room::SendToPlayer(sessionId, pkt) 메서드를 통한 개별 플레이어 안전 전송 구조 마련
- 로비 관리 리팩토링: RoomManager의 로비 세션 관리를 포인터 기반 맵에서 ID 기반 벡터로 전환
- 빌드 이슈 해결: GameEvents.h 중복 포함 방지(pragma once), PlayerFactory 오타 수정 및 RoomManager 벡터 검색 로직 수정

+++ src/System/Dispatcher/MessagePool.cpp
@@ -35,7 +35,7 @@ PacketMessage *MessagePool::AllocatePacket(uint16_t bodySize)
         size_t allocSize = PacketMessage::CalculateAllocSize(bodySize);
         void *block = ::operator new(allocSize);
         PacketMessage *msg = new (block) PacketMessage();
-        msg->type = (uint32_t)MessageType::PACKET;
+        msg->type = MessageType::PACKET;
         msg->length = bodySize;
         msg->isPooled = false;
         return msg;
@@ -47,7 +47,7 @@ PacketMessage *MessagePool::AllocatePacket(uint16_t bodySize)
 
     // Placement New
     PacketMessage *msg = new (block) PacketMessage();
-    msg->type = (uint32_t)MessageType::PACKET; // Ensure consistent type
+    msg->type = MessageType::PACKET; // Ensure consistent type
     msg->length = bodySize;
     msg->isPooled = true;
 
@@ -69,7 +69,7 @@ TimerMessage *MessagePool::AllocateTimer()
     if (!block)
         return nullptr;
     TimerMessage *msg = new (block) TimerMessage();
-    msg->type = static_cast<uint32_t>(MessageType::LOGIC_TIMER);
+    msg->type = MessageType::LOGIC_TIMER;
     // Re-using LOGIC_JOB or adding LOGIC_TIMER in future?
     // Let's check IMessage.h for MessageType enum. It has LOGIC_JOB.
     // Let's stick to LOGIC_JOB or add LOGIC_TIMER?
@@ -84,7 +84,7 @@ TimerExpiredMessage *MessagePool::AllocateTimerExpired()
     if (!block)
         return nullptr;
     TimerExpiredMessage *msg = new (block) TimerExpiredMessage();
-    msg->type = static_cast<uint32_t>(MessageType::LOGIC_TIMER_EXPIRED);
+    msg->type = MessageType::LOGIC_TIMER_EXPIRED;
     return msg;
 }
 
@@ -94,7 +94,7 @@ TimerAddMessage *MessagePool::AllocateTimerAdd()
     if (!block)
         return nullptr;
     TimerAddMessage *msg = new (block) TimerAddMessage();
-    msg->type = static_cast<uint32_t>(MessageType::LOGIC_TIMER_ADD);
+    msg->type = MessageType::LOGIC_TIMER_ADD;
     return msg;
 }
 
@@ -104,7 +104,7 @@ TimerCancelMessage *MessagePool::AllocateTimerCancel()
     if (!block)
         return nullptr;
     TimerCancelMessage *msg = new (block) TimerCancelMessage();
-    msg->type = static_cast<uint32_t>(MessageType::LOGIC_TIMER_CANCEL);
+    msg->type = MessageType::LOGIC_TIMER_CANCEL;
     return msg;
 }
 
@@ -114,7 +114,7 @@ TimerTickMessage *MessagePool::AllocateTimerTick()
     if (!block)
         return nullptr;
     TimerTickMessage *msg = new (block) TimerTickMessage();
-    msg->type = static_cast<uint32_t>(MessageType::LOGIC_TIMER_TICK);
+    msg->type = MessageType::LOGIC_TIMER_TICK;
     return msg;
 }
