+++ src/System/Session/GatewaySession.cpp
@@ -202,7 +202,7 @@ void GatewaySession::OnRecv(size_t bytesTransferred)
             return;
         }
 
-        msg->type = (uint32_t)System::MessageType::NETWORK_DATA;
+        msg->type = System::MessageType::NETWORK_DATA;
         msg->sessionId = _id;
         msg->session = this;
 
@@ -552,7 +552,7 @@ void GatewaySession::OnConnect()
     _connected.store(true, std::memory_order_relaxed);
 
     System::EventMessage *msg = System::MessagePool::AllocateEvent();
-    msg->type = (uint32_t)System::MessageType::NETWORK_CONNECT;
+    msg->type = System::MessageType::NETWORK_CONNECT;
     msg->sessionId = GetId();
     msg->session = this;
     if (_dispatcher)
@@ -571,7 +571,7 @@ void GatewaySession::OnDisconnect()
     if (_connected.exchange(false))
     {
         System::EventMessage *msg = System::MessagePool::AllocateEvent();
-        msg->type = (uint32_t)System::MessageType::NETWORK_DISCONNECT;
+        msg->type = System::MessageType::NETWORK_DISCONNECT;
         msg->sessionId = GetId();
         msg->session = this;
         if (_dispatcher)
